# api-docs.json
{"openapi":"3.0.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/workouts":{"get":{"tags":["workout-controller"],"operationId":"getAllWorkouts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Workout"}}}}}}},"put":{"tags":["workout-controller"],"operationId":"updateWorkout","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Workout"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Workout"}}}}}},"post":{"tags":["workout-controller"],"operationId":"createWorkout","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Workout"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Workout"}}}}}}},"/api/doctors/{id}":{"get":{"tags":["doctor-controller"],"operationId":"getDoctorById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Doctor"}}}}}},"put":{"tags":["doctor-controller"],"operationId":"updateDoctor","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Doctor"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Doctor"}}}}}}},"/api/user/{id}/updateProgress/{userPlanId}":{"post":{"tags":["user-controller"],"operationId":"updateProgress","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"userPlanId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"date","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UpdateProgressResponse"}}}}}}},"/api/plans":{"get":{"tags":["plan-controller"],"operationId":"getAllPlans","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Plan"}}}}}}},"post":{"tags":["plan-controller"],"operationId":"createPlan","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlanCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Plan"}}}}}}},"/api/plans/{planId}/workouts/{workoutId}":{"post":{"tags":["plan-controller"],"operationId":"addWorkout","parameters":[{"name":"planId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"workoutId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Plan"}}}}}},"delete":{"tags":["plan-controller"],"operationId":"removeWorkout","parameters":[{"name":"planId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"workoutId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Plan"}}}}}}},"/api/feedback":{"post":{"tags":["feedback-controller"],"operationId":"submitFeedback","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GeneralResponse"}}}}}}},"/api/doctors/{id}/updateCurrentUserPlan":{"post":{"tags":["doctor-controller"],"operationId":"updateCurrentUserPlan","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocUpdateUserPlan"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GeneralResponse"}}}}}}},"/api/bookings":{"post":{"tags":["booking-controller"],"operationId":"createBooking","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookingRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GeneralResponse"}}}}}}},"/api/auth/patient/signup":{"post":{"tags":["auth-controller"],"operationId":"signupPatient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpRequestPatientDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/doctor/signup":{"post":{"tags":["auth-controller"],"operationId":"signupDoctor","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpRequestDoctorDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/admin/populateWorkouts":{"post":{"tags":["data-population-controller"],"operationId":"populateWorkouts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/user/{id}":{"get":{"tags":["user-controller"],"operationId":"getUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"patch":{"tags":["user-controller"],"operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/workouts/{id}":{"get":{"tags":["workout-controller"],"operationId":"getWorkoutById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Workout"}}}}}},"delete":{"tags":["workout-controller"],"operationId":"deleteWorkout","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/workouts/search":{"get":{"tags":["workout-controller"],"operationId":"searchByName","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Workout"}}}}}}}},"/api/user/{id}/planHistory":{"get":{"tags":["user-controller"],"operationId":"getPatientPlanHistory","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserPlanHistoryResponse"}}}}}}},"/api/user/{id}/graphData":{"get":{"tags":["user-controller"],"operationId":"getGraphData","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"startDate","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"endDate","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TimeSeriesGraphDto"}}}}}}},"/api/user/{id}/currentPlanAndProgress":{"get":{"tags":["user-controller"],"operationId":"getCurrentPlanAndProgress","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CurrentPlanAndProgressResponse"}}}}}}},"/api/plans/{id}":{"get":{"tags":["plan-controller"],"operationId":"getPlan","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Plan"}}}}}},"delete":{"tags":["plan-controller"],"operationId":"deletePlan","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/feedback/userPlan/{userPlanId}":{"get":{"tags":["feedback-controller"],"operationId":"getLatestFeedbackByUserPlanId","parameters":[{"name":"userPlanId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FeedbackResponseDto"}}}}}}},"/api/doctors":{"get":{"tags":["doctor-controller"],"operationId":"getAllDoctors","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Doctor"}}}}}}}},"/api/doctors/specialization/{specialization}":{"get":{"tags":["doctor-controller"],"operationId":"getDoctorsBySpecialization","parameters":[{"name":"specialization","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Doctor"}}}}}}}},"/api/doctors/search/{name}":{"get":{"tags":["doctor-controller"],"operationId":"searchDoctors","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Doctor"}}}}}}}},"/api/bookings/doctor/{doctorId}/availabilities":{"get":{"tags":["booking-controller"],"operationId":"getDoctorBookingAvailabilites","parameters":[{"name":"doctorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"startDate","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"endDate","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BookingAvailabilitiesResponse"}}}}}}},"/api/auth/profile":{"get":{"tags":["auth-controller"],"operationId":"login_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDetailsDto"}}}}}}},"/api/admin/clearAllData":{"get":{"tags":["data-population-controller"],"operationId":"clearDatabase","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"Workout":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string","maxLength":100,"minLength":0},"description":{"type":"string","maxLength":500,"minLength":0},"imageUrl":{"type":"string"},"videoUrl":{"type":"string"},"difficulty":{"type":"string"},"duration":{"type":"integer","format":"int32"},"patientNote":{"type":"string","maxLength":1000,"minLength":0},"alternateExerciseId":{"type":"integer","format":"int64"},"sets":{"type":"integer","format":"int32"},"reps":{"type":"integer","format":"int32"}},"required":["name"]},"Doctor":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["PATIENT","DOCTOR"]},"firstName":{"type":"string"},"lastName":{"type":"string"},"phone":{"type":"string"},"specialization":{"type":"string"},"emergencyContact":{"type":"string"},"availabilities":{"type":"array","items":{"$ref":"#/components/schemas/DoctorAvailability"}}}},"DoctorAvailability":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"dayOfWeek":{"type":"string","enum":["MONDAY","TUESDAY","WEDNESDAY","THURSDAY","FRIDAY","SATURDAY","SUNDAY"]},"startTime":{"type":"string"},"endTime":{"type":"string"}}},"UpdateProgressResponse":{"type":"object","properties":{"status":{"type":"string","enum":["SUCCESS","FAIL","CONFIRMED","PENDING","REJECTED","IN_PROGRESS","COMPLETED"]},"totalCount":{"type":"integer","format":"int32"},"completedCount":{"type":"integer","format":"int32"},"message":{"type":"string"}}},"PlanCreateRequest":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"duration":{"type":"integer","format":"int32"},"workoutIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"Plan":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"introduction":{"type":"string"},"duration":{"type":"integer","format":"int32"},"workouts":{"type":"array","items":{"$ref":"#/components/schemas/Workout"}}}},"FeedbackDto":{"type":"object","properties":{"userPlanId":{"type":"integer","format":"int32"},"feedback":{"type":"string"},"exerciseHelpFulness":{"type":"integer","format":"int32"},"exerciseEffectiveness":{"type":"integer","format":"int32"},"howWellDidyouFare":{"type":"integer","format":"int32"},"suggestedImprovement":{"type":"string"},"overallRating":{"type":"integer","format":"int32"}}},"GeneralResponse":{"type":"object","properties":{"status":{"type":"string","enum":["SUCCESS","FAIL","CONFIRMED","PENDING","REJECTED","IN_PROGRESS","COMPLETED"]},"message":{"type":"string"},"errors":{"type":"object","additionalProperties":{"type":"string"}}}},"DocUpdateUserPlan":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"planName":{"type":"string"},"doctorNote":{"type":"string"},"workoutIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"CreateBookingRequest":{"type":"object","properties":{"doctorId":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"date":{"type":"string","format":"date"},"doctorAvailabilityId":{"type":"integer","format":"int64"},"remark":{"type":"string"}},"required":["date","doctorAvailabilityId","doctorId","userId"]},"MedicalRecordsDto":{"type":"object","properties":{"lvef":{"type":"integer","format":"int32","maximum":100,"minimum":0},"nyhaClass":{"type":"string"},"hypertensionUpper":{"type":"integer","format":"int32","maximum":250,"minimum":50},"hypertensionLower":{"type":"integer","format":"int32","maximum":150,"minimum":30},"cholesterolLevel":{"type":"number","format":"double","minimum":0},"oxygenSaturation":{"type":"integer","format":"int32","maximum":100,"minimum":50},"smokingHistory":{"type":"string"},"diabetes":{"type":"number","format":"double","maximum":20,"minimum":0},"exerciseTolerance":{"type":"number","format":"double","minimum":0},"cardiacEventType":{"type":"string"},"dateOfLastCardiacEvent":{"type":"string","format":"date"},"alcoholIntake":{"type":"string"},"physicalActivityBeforeEvent":{"type":"string"},"cardiacError":{"type":"array","items":{"type":"string"}}}},"SignUpRequestPatientDto":{"type":"object","properties":{"email":{"type":"string","minLength":1},"password":{"type":"string","maxLength":2147483647,"minLength":6},"firstName":{"type":"string","minLength":1},"lastName":{"type":"string","minLength":1},"phone":{"type":"string","minLength":1,"pattern":"^[0-9]{10,15}$"},"age":{"type":"integer","format":"int32"},"height":{"type":"number","format":"float"},"weight":{"type":"number","format":"float"},"medicalRecords":{"$ref":"#/components/schemas/MedicalRecordsDto"}},"required":["age","email","firstName","height","lastName","medicalRecords","password","phone","weight"]},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"DoctorAvailabilityDto":{"type":"object","properties":{"dayOfWeek":{"type":"string","enum":["MONDAY","TUESDAY","WEDNESDAY","THURSDAY","FRIDAY","SATURDAY","SUNDAY"]},"startTime":{"type":"string"},"endTime":{"type":"string"}}},"SignUpRequestDoctorDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"specialization":{"type":"string"},"emergencyContact":{"type":"string"},"availabilities":{"type":"array","items":{"$ref":"#/components/schemas/DoctorAvailabilityDto"}}}},"MedicalRecords":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"lvef":{"type":"integer","format":"int32","maximum":100,"minimum":0},"nyhaClass":{"type":"string"},"hypertensionUpper":{"type":"integer","format":"int32","maximum":250,"minimum":50},"hypertensionLower":{"type":"integer","format":"int32","maximum":150,"minimum":30},"cholesterolLevel":{"type":"number","format":"double","minimum":0},"oxygenSaturation":{"type":"integer","format":"int32","maximum":100,"minimum":50},"smokingHistory":{"type":"string"},"diabetes":{"type":"number","format":"double","maximum":20,"minimum":0},"exerciseTolerance":{"type":"number","format":"double","minimum":0},"cardiacEventType":{"type":"string"},"dateOfLastCardiacEvent":{"type":"string","format":"date"},"alcoholIntake":{"type":"string"},"physicalActivityBeforeEvent":{"type":"string"},"cardiacError":{"type":"array","items":{"type":"string"}}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["PATIENT","DOCTOR"]},"firstName":{"type":"string"},"lastName":{"type":"string"},"phoneNumber":{"type":"string"},"age":{"type":"integer","format":"int32","maximum":120,"minimum":1},"height":{"type":"number","format":"float"},"weight":{"type":"number","format":"float"},"medicalRecords":{"$ref":"#/components/schemas/MedicalRecords"}}},"UserPlanHistoryResponse":{"type":"object","properties":{"userPlans":{"type":"array","items":{"$ref":"#/components/schemas/UserPlans"}}}},"UserPlans":{"type":"object","properties":{"userPlanId":{"type":"integer","format":"int32"},"planName":{"type":"string"},"planDescription":{"type":"string"},"planNotes":{"type":"string"},"createdDateTime":{"type":"string","format":"date-time"},"updatedBy":{"type":"string"},"workouts":{"type":"array","items":{"$ref":"#/components/schemas/WorkoutDto"}},"current":{"type":"boolean"}}},"WorkoutDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"imageUrl":{"type":"string"},"videoUrl":{"type":"string"},"difficulty":{"type":"string"},"duration":{"type":"integer","format":"int32"},"patientNote":{"type":"string"},"alternateExerciseId":{"type":"integer","format":"int64"},"sets":{"type":"integer","format":"int32"},"reps":{"type":"integer","format":"int32"}}},"TimeSeriesGraphDto":{"type":"object","properties":{"status":{"type":"string","enum":["SUCCESS","FAIL","CONFIRMED","PENDING","REJECTED","IN_PROGRESS","COMPLETED"]},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/XYData"}}}},"XYData":{"type":"object","properties":{"date":{"type":"string","format":"date"},"percentage":{"type":"number","format":"double"}}},"CurrentPlanAndProgressResponse":{"type":"object","properties":{"status":{"type":"string","enum":["SUCCESS","FAIL","CONFIRMED","PENDING","REJECTED","IN_PROGRESS","COMPLETED"]},"message":{"type":"string"},"userPlanId":{"type":"integer","format":"int32"},"planName":{"type":"string"},"planDescription":{"type":"string"},"doctorNotes":{"type":"string"},"modifiedDoctor":{"type":"string"},"completed":{"type":"integer","format":"int32"},"total":{"type":"integer","format":"int32"},"progressPercentage":{"type":"number","format":"double"},"todaysDate":{"type":"string","format":"date"},"workouts":{"type":"array","items":{"$ref":"#/components/schemas/WorkoutDto"}}}},"Feedback":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userPlanId":{"type":"integer","format":"int32"},"feedback":{"type":"string"},"exerciseHelpFulness":{"type":"integer","format":"int32"},"exerciseEffectiveness":{"type":"integer","format":"int32"},"howWellDidyouFare":{"type":"integer","format":"int32"},"suggestedImprovement":{"type":"string"},"overallRating":{"type":"integer","format":"int32"}}},"FeedbackResponseDto":{"type":"object","properties":{"status":{"type":"string","enum":["SUCCESS","FAIL","CONFIRMED","PENDING","REJECTED","IN_PROGRESS","COMPLETED"]},"message":{"type":"string"},"feedback":{"$ref":"#/components/schemas/Feedback"}}},"Availability":{"type":"object","properties":{"availabilityId":{"type":"integer","format":"int64"},"startTime":{"type":"string"},"endTime":{"type":"string"}}},"BookingAvailabilitiesResponse":{"type":"object","properties":{"slots":{"type":"array","items":{"$ref":"#/components/schemas/Slot"}}}},"Slot":{"type":"object","properties":{"date":{"type":"string","format":"date"},"dayOfWeek":{"type":"string","enum":["MONDAY","TUESDAY","WEDNESDAY","THURSDAY","FRIDAY","SATURDAY","SUNDAY"]},"availabilities":{"type":"array","items":{"$ref":"#/components/schemas/Availability"}}}},"UserDetailsDto":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["PATIENT","DOCTOR"]}}}}}}