# src/main/java/com/dexter/fyp/backend/dto/CreateBookingRequest.java
package com.dexter.fyp.backend.dto;

import jakarta.validation.constraints.Future;
import jakarta.validation.constraints.NotNull;
import lombok.Data;

import java.time.LocalDate;

@Data
public class CreateBookingRequest {

    @NotNull(message = "Doctor ID is required")
    private Long doctorId;

    @NotNull(message = "User ID is required")
    private Long userId;

    @NotNull(message = "Date is required")
    @Future(message = "Date must be in the future")
    private LocalDate date;

    @NotNull(message = "Doctor Availability ID is required")
    private Long doctorAvailabilityId;

    private String remark;
}